requirements.txt
acres==0.3.0
aiofiles==24.1.0
anyio==4.9.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
colorama==0.4.6
comtypes==1.4.10
configobj==5.0.9
configparser==7.2.0
docx2pdf==0.1.8
easyocr==1.7.2
etelemetry==0.3.1
exceptiongroup==1.2.2
filelock==3.18.0
Flask==3.1.0
frontend==0.0.3
fsspec==2025.3.2
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.69.2
h11==0.14.0
httplib2==0.22.0
idna==3.10
imageio==2.37.0
importlib_resources==6.5.2
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
lazy_loader==0.4
looseversion==1.3.0
lxml==5.3.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
nibabel==5.3.2
ninja==1.11.1.4
nipype==1.10.0
numpy==2.2.4
oauthlib==3.2.2
opencv-python-headless==4.11.0.86
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pillow==11.1.0
proto-plus==1.26.1
protobuf==6.30.2
prov==2.0.1
puremagic==1.28
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pydot==3.0.4
PyMuPDF==1.25.5
pyparsing==3.2.3
PyPDF2==3.0.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-pptx==1.0.2
pytils==0.4.3
pytz==2025.2
pywin32==310
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==6.3.2
reportlab==4.3.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scikit-image==0.25.2
scipy==1.15.2
shapely==2.1.0
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
sympy==1.13.1
tifffile==2025.3.30
tools==0.1.9
tqdm==4.67.1
traits==7.0.2
typing_extensions==4.13.1
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
Werkzeug==3.1.3
XlsxWriter==3.2.2
